+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## _(in-progress)_
+* New: Add new setting "More filters" to allow for user-specified filters to be processed for text replacements
 * Change: Update plugin framework to 061
     * 061:
     * Fix bug preventing settings from getting saved